requirements.txt
requests
lxml
mysql-connector-python
